move search infrastructure to ADF core
