+++ lib/core/index.ts
@@ -52,6 +52,7 @@ export * from './collapsable';
 export * from './card-view';
 export * from './app-config';
 export * from './form';
+export * from './search';
 
 export * from './pipes';
 export * from './services';
