+++ lib/core/search/public-api.ts
@@ -1,0 +1,29 @@
+/*!
+ * @license
+ * Copyright 2016 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export { FacetFieldBucket } from './facet-field-bucket.interface';
+export { FacetField } from './facet-field.interface';
+export { FacetQuery } from './facet-query.interface';
+export { FilterQuery } from './filter-query.interface';
+export { ResponseFacetField } from './response-facet-field.interface';
+export { ResponseFacetQuery } from './response-facet-query.interface';
+export { SearchCategory } from './search-category.interface';
+export { SearchComponentSettings } from './search-component-settings.interface';
+export { SearchComponent } from './search-component.interface';
+export { SearchConfiguration } from './search-configuration.interface';
+export { SearchQueryBuilder } from './search-query-builder';
+export { SearchRange } from './search-range.interface';
