+++ lib/core/search/search-category.interface.ts
@@ -1,0 +1,29 @@
+/*!
+ * @license
+ * Copyright 2016 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { SearchComponentSettings } from "./search-component-settings.interface";
+
+export interface SearchCategory {
+    id: string;
+    name: string;
+    enabled: boolean;
+    expanded: boolean;
+    component: {
+        selector: string;
+        settings: SearchComponentSettings;
+    };
+}
